+++ packages/hydrogen/src/components/CartProvider/tests/fixtures.ts
@@ -1,5 +1,6 @@
-import {CART_LINE} from '../../../components/CartLineProvider/tests/fixtures';
-import {getPrice} from '../../../../src/utilities/tests/price';
+import {CART_LINE} from '../../CartLineProvider/tests/fixtures';
+import {getPrice} from '../../../utilities/tests/price';
+import {flattenConnection} from '../../../utilities';
 
 export const CART = {
   id: 'abc',
@@ -19,4 +20,9 @@ export const CART = {
 export const CART_WITH_LINES = {
   ...CART,
   lines: {edges: [{node: CART_LINE}]},
+};
+
+export const CART_WITH_LINES_FLATTENED = {
+  ...CART,
+  lines: flattenConnection(CART_WITH_LINES.lines),
 };